requirements.txt
# Use the PyTorch Nightly index for CUDA 12.8
--extra-index-url https://download.pytorch.org/whl/nightly/cu128

# Core AI stack (pinned to prevent deep resolution loops)
torch
torchvision
torchaudio
transformers>=4.48.0
accelerate>=1.1.0
bitsandbytes>=0.45.0
diffusers>=0.31.0

# Foundation & Utilities
numpy
requests
sentencepiece
tokenizers
uvicorn
httpx
scikit-learn
einops
einops-exts
timm
openai-clip
fsspec
kornia
matplotlib
ninja
omegaconf
open-clip-torch==2.24.0 # new versions do not work
opencv-python
pandas
Pillow
pytorch-lightning
PyYAML
scipy
tqdm
triton
urllib3
webdataset
facexlib
k-diffusion
huggingface_hub